proof of concept
After I uploaded the SVG I just couldn't help myself and tried to bring it to life with a little javascript. It does basically what I wanted it to, but only the first two months have IDs assigned so it's not fully functional. There is a small interface if you hover up in the left hand corner of the browser window.